added wall to build

+++ include/data/Code.h
@@ -14,20 +14,19 @@ class Jitable;
 
 class Code : public Data {
 private:
-    int start, end, premature_end;
+    size_t start, end, premature_end;
     const std::vector<Command>* program;
 
 public:
     bool scheduleForParallelExecution;
     Jitable* jitable;
     
-    explicit Code(const std::vector<Command>* programAt, int startAt, int endAt, int premature_end);
-    
+    explicit Code(const std::vector<Command>* programAt, size_t startAt, size_t endAt, size_t premature_end);
     Code* copy() const {Code* ret = new Code(program, start, end, premature_end);ret->jitable=jitable; return ret;}
     std::string toString(BMemory* memory)override;
-    int getStart() const;
-    int getEnd() const;
-    int getOptimizedEnd() const;
+    size_t getStart() const;
+    size_t getEnd() const;
+    size_t getOptimizedEnd() const;
     const std::vector<Command>* getProgram() const;
 };
